test: split test_execution_modes.py under the 800-line file-size convention
Splits tests/integration/bus/test_execution_modes.py (808 lines) into
test_execution_modes.py (core overlap-mode dispatch behavior) and
test_execution_modes_guards.py (guard internals: live-execution-count
snapshots, persisted backpressure policy, the stall watchdog, and queued
pending_done draining), following the existing test_bus_error_handler.py /
test_bus_error_handler_combos.py precedent in the same directory.

Shared helpers (ENTITY, EVENT_LOOP_YIELDS, pump_event_loop, fire) move into
tests/integration/bus/helpers.py so both files can import them.

Pure test reorganization — no behavior changes.

Closes #1584
