+++ tests/integration/bus/CLAUDE.md
@@ -7,6 +7,7 @@
 ## Shared helpers
 
 - `from hassette.test_utils import HassetteHarness` — builder used directly by `bus_harness`; seed state via `harness.seed_state()`
+- This directory's `helpers.py` also exposes `seed`, `fire`, `pump_event_loop`, `ENTITY`, and `EVENT_LOOP_YIELDS` — shared by `test_execution_modes.py` and `test_execution_modes_guards.py` for overlap-mode dispatch tests.
 
 ## Key conventions
